allow extract blks
dev only

+++ src/Snap.Hutao/Snap.Hutao/Service/Game/Package/Advanced/GameAssetOperation.cs
@@ -147,7 +147,7 @@ internal abstract partial class GameAssetOperation : IGameAssetOperation
 
         foreach (AssetProperty asset in assets)
         {
-            string assetPath = Path.Combine(context.Operation.GameFileSystem.GameDirectory, asset.AssetName);
+            string assetPath = Path.Combine(context.Operation.GameFileSystem.ExtractDirectory, asset.AssetName);
 
             if (asset.AssetType is 64)
             {
@@ -287,7 +287,7 @@ internal abstract partial class GameAssetOperation : IGameAssetOperation
                                 }
                             }
 
-                            if (!context.DuplicatedChunkNames.ContainsKey(chunk.ChunkName))
+                            if (context.Operation.Kind is GamePackageOperationKind.Update && !context.DuplicatedChunkNames.ContainsKey(chunk.ChunkName))
                             {
                                 FileOperation.Delete(chunkPath);
                             }
@@ -320,7 +320,16 @@ internal abstract partial class GameAssetOperation : IGameAssetOperation
                 }
             }
 
-            string newAssetPath = Path.Combine(context.Operation.GameFileSystem.GameDirectory, asset.NewAsset.AssetName);
+            string assetName = asset.NewAsset.AssetName;
+            if (context.Operation.Kind is GamePackageOperationKind.Extract)
+            {
+                assetName = Path.GetFileName(assetName);
+            }
+
+            string newAssetPath = Path.Combine(context.Operation.GameFileSystem.ExtractDirectory, assetName);
+            string? directory = Path.GetDirectoryName(newAssetPath);
+            ArgumentNullException.ThrowIfNull(directory);
+            Directory.CreateDirectory(directory);
             using (FileStream newAssetFileStream = File.Create(newAssetPath))
             {
                 newAssetStream.Position = 0;
